% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aqs_flags}
\alias{aqs_flags}
\title{EPA AQS qualifier flag definitions}
\format{
An object of class \code{data.frame} with 163 rows and 7 columns.
}
\usage{
aqs_flags
}
\description{
This dataset was downloaded from the AQS website. An additional column with
the corresponding NARSTO flag was added by the ASRC.
}
\keyword{datasets}
