% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clock_audits}
\alias{clock_audits}
\title{Clock audit results}
\format{An object of class \code{data.frame} with 12 rows and 5 columns.}
\usage{
clock_audits
}
\description{
Clock audit results
}
\keyword{datasets}
