% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sources}
\alias{data_sources}
\title{Data sources}
\format{
An object of class \code{data.frame} with 9 rows and 3 columns.
}
\usage{
data_sources
}
\description{
Data sources
}
\keyword{datasets}
