% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{estimate_cals}
\alias{estimate_cals}
\title{Estimate calibration values}
\usage{
estimate_cals(x, y, k, xout, breaks)
}
\arguments{
\item{x}{Calibration check times (POSIXct).}

\item{y}{Calibration check values.}

\item{k}{Median filter window width. If \code{NA} no smoothing is applied.}

\item{xout}{Times at which to estimate the calibration value.}

\item{breaks}{Times of discontinuities caused by instrument adjustments.}
}
\value{
Estimated calibration values at times \code{xout}.
}
\description{
Estimate calibration values by applying a median filter and interpolating
between the filtered values.
}
