% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_nysatmoschem}
\alias{etl_extract.etl_nysatmoschem}
\title{My ETL functions}
\usage{
\method{etl_extract}{etl_nysatmoschem}(obj, user, password, sites = NULL,
  data_sources = NULL, years = NULL, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}
}
\description{
My ETL functions
}
\examples{
\dontrun{
if (require(dplyr)) {
  obj <- etl("nysatmoschem") \%>\%
    etl_create()
}
}
}
