% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gilibrator}
\alias{gilibrator}
\title{Gilibrator flow measurements}
\format{
An object of class \code{data.frame} with 58 rows and 6 columns.
}
\usage{
gilibrator
}
\description{
Gilibrator flow measurements
}
\keyword{datasets}
