% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{instruments}
\alias{instruments}
\title{Instruments}
\format{An object of class \code{data.frame} with 26 rows and 6 columns.}
\usage{
instruments
}
\description{
Instruments
}
\keyword{datasets}
