% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measurement_sources}
\alias{measurement_sources}
\title{Measurement instruments}
\format{An object of class \code{data.frame} with 25 rows and 5 columns.}
\usage{
measurement_sources
}
\description{
Measurement instruments
}
\keyword{datasets}
