% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measurement_types}
\alias{measurement_types}
\title{Information and processing instructions for each type of
measurement}
\format{
An object of class \code{data.frame} with 83 rows and 22 columns.
}
\usage{
measurement_types
}
\description{
Information and processing instructions for each type of
measurement
}
\keyword{datasets}
