% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{narsto_agg_flag}
\alias{narsto_agg_flag}
\title{Flag aggregated data}
\usage{
narsto_agg_flag(p, below_mdl = FALSE)
}
\arguments{
\item{p}{Proportion of valid measurements.}

\item{below_mdl}{Whether the aggregated value is below the method detection
limit. MDLs for aggregates are generally lower than the limits for
individual measurements.}
}
\value{
A NARSTO flag.
}
\description{
Return the NARSTO flag for an aggregated measurement (for example, hourly
average or maximum). See
\insertCite{@ @narsto_consensus_2001, p. 10;textual}{atmoschem.process}.
}
\examples{
narsto_agg_flag(.7, below_mdl = FALSE)

}
\references{
\insertAllCited{}
}
