% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{narsto_flags}
\alias{narsto_flags}
\title{NARSTO flag definitions}
\format{
An object of class \code{data.frame} with 11 rows and 4 columns.
}
\usage{
narsto_flags
}
\description{
Flag definitions used by NARSTO, as described in
\insertCite{christensen_narsto_1999;textual}{atmoschem.process} and
\insertCite{hook_developing_2005;textual}{atmoschem.process}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
