% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asrc_calibrations.R
\name{read_pdf_form}
\alias{read_pdf_form}
\title{Get fields from a pdf form}
\usage{
read_pdf_form(f)
}
\arguments{
\item{f}{Path to pdf file.}
}
\value{
A list of the field values.
}
\description{
Get the field values from a pdf form (AcroForms). This function is a
streamlined alternative to \code{staplr::get_fields} that's about twice as
fast.
}
\examples{
# read the staplr example file
pdf = system.file('testForm.pdf', package = 'staplr')
head(read_pdf_form(pdf))

}
\seealso{
\link[staplr]{get_fields}
}
