% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{report_columns}
\alias{report_columns}
\title{Report column information}
\format{An object of class \code{data.frame} with 26 rows and 6 columns.}
\usage{
report_columns
}
\description{
Report column information
}
\keyword{datasets}
