% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aqs_flags}
\alias{aqs_flags}
\title{AQS data qualifier flag definitions}
\format{
An object of class \code{data.frame} with 163 rows and 7 columns.
}
\usage{
aqs_flags
}
\description{
Flag definitions used by the EPA's Air Quality System
\insertCite{us_environmental_protection_agency_air_nodate}{nysatmoschem}, as
provided by the EPA. An additional column with the corresponding NARSTO flags
was added by the ASRC.
}
\references{
\insertAllCited{}
}
\seealso{
\url{https://aqs.epa.gov/aqsweb/documents/codetables/qualifiers.html}
}
\keyword{datasets}
