% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{autocals}
\alias{autocals}
\title{The instrument auto-calibration schedule}
\format{An object of class \code{data.frame} with 40 rows and 7 columns.}
\usage{
autocals
}
\description{
The instrument auto-calibration schedule
}
\keyword{datasets}
