% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cal_flags}
\alias{cal_flags}
\title{Calibration flags}
\format{An object of class \code{data.frame} with 15 rows and 6 columns.}
\usage{
cal_flags
}
\description{
Calibration flags
}
\keyword{datasets}
