% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{narsto_flags}
\alias{narsto_flags}
\title{NARSTO flag definitions taken from
"Developing Data Management Policy and Guidance Documents for your NARSTO Program or Project"}
\format{
An object of class \code{data.frame} with 11 rows and 4 columns.
}
\usage{
narsto_flags
}
\description{
NARSTO flag definitions taken from
"Developing Data Management Policy and Guidance Documents for your NARSTO Program or Project"
}
\keyword{datasets}
